backend/requirements.txt
fastapi
uvicorn[standard]
motor>=3.3,<3.4
pymongo>=4.6,<4.8
beanie>=1.25,<2.0
python-jose[cryptography]
passlib[bcrypt]
openai
python-dotenv
apscheduler
httpx
beautifulsoup4
Pillow
python-multipart
aiofiles
